factor out statement search handlers
